+++ kernels/geometry/curveNi_intersector.h
@@ -148,7 +148,7 @@ namespace embree
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
           
           unsigned int vertexID = geom->curve(primID);
-          Vec3fa a0,a1,a2,a3,n0,n1; geom->gather(a0,a1,a2,a3,n0,n1,vertexID);
+          Vec3fa a0,a1,a2,a3,n0,n1,n2,n3; geom->gather(a0,a1,a2,a3,n0,n1,n2,n3,vertexID);
 
           size_t mask1 = mask;
           const size_t i1 = bscf(mask1);
@@ -162,7 +162,7 @@ namespace embree
             }
           }
           
-          Intersector().intersect(pre,ray,geom,primID,a0,a1,a2,a3,n0,n1,Epilog(ray,context,geomID,primID));
+          Intersector().intersect(pre,ray,geom,primID,a0,a1,a2,a3,n0,n1,n2,n3,Epilog(ray,context,geomID,primID));
           mask &= movemask(tNear <= vfloat<M>(ray.tfar));
         }
       }
@@ -184,7 +184,7 @@ namespace embree
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
 
           unsigned int vertexID = geom->curve(primID);
-          Vec3fa a0,a1,a2,a3,n0,n1; geom->gather(a0,a1,a2,a3,n0,n1,vertexID);
+          Vec3fa a0,a1,a2,a3,n0,n1,n2,n3; geom->gather(a0,a1,a2,a3,n0,n1,n2,n3,vertexID);
 
           size_t mask1 = mask;
           const size_t i1 = bscf(mask1);
@@ -198,7 +198,7 @@ namespace embree
             }
           }
 
-          if (Intersector().intersect(pre,ray,geom,primID,a0,a1,a2,a3,n0,n1,Epilog(ray,context,geomID,primID)))
+          if (Intersector().intersect(pre,ray,geom,primID,a0,a1,a2,a3,n0,n1,n2,n3,Epilog(ray,context,geomID,primID)))
             return true;
           
           mask &= movemask(tNear <= vfloat<M>(ray.tfar));
@@ -266,8 +266,8 @@ namespace embree
           const unsigned int geomID = prim.geomID(N);
           const unsigned int primID = prim.primID(N)[i];
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
-          Vec3fa p0,t0,n0,p1,t1,n1; geom->gather_hermite(p0,t0,n0,p1,t1,n1,geom->curve(primID));
-          Intersector().intersect(pre,ray,geom,primID,p0,t0,n0,p1,t1,n1,Epilog(ray,context,geomID,primID));
+          Vec3fa p0,t0,n0,dn0,p1,t1,n1,dn1; geom->gather_hermite(p0,t0,n0,dn0,p1,t1,n1,dn1,geom->curve(primID));
+          Intersector().intersect(pre,ray,geom,primID,p0,t0,p1,t1,n0,dn0,n1,dn1,Epilog(ray,context,geomID,primID));
           mask &= movemask(tNear <= vfloat<M>(ray.tfar));
         }
       }
@@ -287,8 +287,8 @@ namespace embree
           const unsigned int geomID = prim.geomID(N);
           const unsigned int primID = prim.primID(N)[i];
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
-          Vec3fa p0,t0,n0,p1,t1,n1; geom->gather_hermite(p0,t0,n0,p1,t1,n1,geom->curve(primID));
-          if (Intersector().intersect(pre,ray,geom,primID,p0,t0,n0,p1,t1,n1,Epilog(ray,context,geomID,primID)))
+          Vec3fa p0,t0,n0,dn0,p1,t1,n1,dn1; geom->gather_hermite(p0,t0,n0,dn0,p1,t1,n1,dn1,geom->curve(primID));
+          if (Intersector().intersect(pre,ray,geom,primID,p0,t0,p1,t1,n0,dn0,n1,dn1,Epilog(ray,context,geomID,primID)))
             return true;
           
           mask &= movemask(tNear <= vfloat<M>(ray.tfar));
@@ -426,7 +426,7 @@ namespace embree
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
 
           unsigned int vertexID = geom->curve(primID);
-          Vec3fa a0,a1,a2,a3,n0,n1; geom->gather(a0,a1,a2,a3,n0,n1,vertexID);
+          Vec3fa a0,a1,a2,a3,n0,n1,n2,n3; geom->gather(a0,a1,a2,a3,n0,n1,n2,n3,vertexID);
 
           size_t mask1 = mask;
           const size_t i1 = bscf(mask1);
@@ -440,7 +440,7 @@ namespace embree
             }
           }
 
-          Intersector().intersect(pre,ray,k,geom,primID,a0,a1,a2,a3,n0,n1,Epilog(ray,k,context,geomID,primID));
+          Intersector().intersect(pre,ray,k,geom,primID,a0,a1,a2,a3,n0,n1,n2,n3,Epilog(ray,k,context,geomID,primID));
           mask &= movemask(tNear <= vfloat<M>(ray.tfar[k]));
         }
       }
@@ -462,7 +462,7 @@ namespace embree
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
 
           unsigned int vertexID = geom->curve(primID);
-          Vec3fa a0,a1,a2,a3,n0,n1; geom->gather(a0,a1,a2,a3,n0,n1,vertexID);
+          Vec3fa a0,a1,a2,a3,n0,n1,n2,n3; geom->gather(a0,a1,a2,a3,n0,n1,n2,n3,vertexID);
 
           size_t mask1 = mask;
           const size_t i1 = bscf(mask1);
@@ -476,7 +476,7 @@ namespace embree
             }
           }
 
-          if (Intersector().intersect(pre,ray,k,geom,primID,a0,a1,a2,a3,n0,n1,Epilog(ray,k,context,geomID,primID)))
+          if (Intersector().intersect(pre,ray,k,geom,primID,a0,a1,a2,a3,n0,n1,n2,n3,Epilog(ray,k,context,geomID,primID)))
             return true;
           
           mask &= movemask(tNear <= vfloat<M>(ray.tfar[k]));
@@ -544,8 +544,8 @@ namespace embree
           const unsigned int geomID = prim.geomID(N);
           const unsigned int primID = prim.primID(N)[i];
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
-          Vec3fa p0,t0,n0,p1,t1,n1; geom->gather_hermite(p0,t0,n0,p1,t1,n1,geom->curve(primID));
-          Intersector().intersect(pre,ray,k,geom,primID,p0,t0,n0,p1,t1,n1,Epilog(ray,k,context,geomID,primID));
+          Vec3fa p0,t0,n0,dn0,p1,t1,n1,dn1; geom->gather_hermite(p0,t0,n0,dn0,p1,t1,n1,dn1,geom->curve(primID));
+          Intersector().intersect(pre,ray,k,geom,primID,p0,t0,p1,t1,n0,dn0,n1,dn1,Epilog(ray,k,context,geomID,primID));
           mask &= movemask(tNear <= vfloat<M>(ray.tfar[k]));
         }
       }
@@ -565,8 +565,8 @@ namespace embree
           const unsigned int geomID = prim.geomID(N);
           const unsigned int primID = prim.primID(N)[i];
           const CurveGeometry* geom = context->scene->get<CurveGeometry>(geomID);
-          Vec3fa p0,t0,n0,p1,t1,n1; geom->gather_hermite(p0,t0,n0,p1,t1,n1,geom->curve(primID));
-          if (Intersector().intersect(pre,ray,k,geom,primID,p0,t0,n0,p1,t1,n1,Epilog(ray,k,context,geomID,primID)))
+          Vec3fa p0,t0,n0,dn0,p1,t1,n1,dn1; geom->gather_hermite(p0,t0,n0,dn0,p1,t1,n1,dn1,geom->curve(primID));
+          if (Intersector().intersect(pre,ray,k,geom,primID,p0,t0,p1,t1,n0,dn0,n1,dn1,Epilog(ray,k,context,geomID,primID)))
             return true;
           
           mask &= movemask(tNear <= vfloat<M>(ray.tfar[k]));