+++ tutorials/common/scenegraph/scenegraph.h
@@ -956,6 +956,7 @@ namespace embree
         positions(transformMSMBlurBuffer(imesh->positions,spaces)),
         normals(transformMSMBlurNormalBuffer(imesh->normals,spaces)),
         tangents(transformMSMBlurVectorBuffer(imesh->tangents,spaces)),
+        dnormals(transformMSMBlurVectorBuffer(imesh->dnormals,spaces)),
         hairs(imesh->hairs), flags(imesh->flags), material(imesh->material), tessellation_rate(imesh->tessellation_rate) {}
 
       virtual void setMaterial(Ref<MaterialNode> material) {
@@ -1016,6 +1017,7 @@ namespace embree
       std::vector<avector<Vertex>> positions; //!< hair control points (x,y,z,r) for multiple timesteps
       std::vector<avector<Vertex>> normals;   //!< hair control normals (nx,ny,nz) for multiple timesteps
       std::vector<avector<Vertex>> tangents;  //!< hair control tangents (tx,ty,tz,tr) for multiple timesteps
+      std::vector<avector<Vertex>> dnormals;  //!< hair control normal derivatives (nx,ny,nz) for multiple timesteps
       std::vector<Hair> hairs;                //!< list of hairs
       std::vector<unsigned char> flags;       //!< left, right end cap flags