+++ tutorials/common/tutorial/scene_device.cpp
@@ -264,7 +264,7 @@ namespace embree
   }
   
   ISPCHairSet::ISPCHairSet (TutorialScene* scene_in, RTCGeometryType type, Ref<SceneGraph::HairSetNode> in)
-    : geom(CURVES), normals(nullptr), tangents(nullptr), type(type)
+    : geom(CURVES), normals(nullptr), tangents(nullptr), dnormals(nullptr), type(type)
   {
     positions = new Vec3fa*[in->numTimeSteps()];
     for (size_t i=0; i<in->numTimeSteps(); i++)
@@ -280,6 +280,12 @@ namespace embree
       tangents = new Vec3fa*[in->numTimeSteps()];
       for (size_t i=0; i<in->numTimeSteps(); i++)
         tangents[i] = in->tangents[i].data();
+    }
+
+    if (in->dnormals.size()) {
+      dnormals = new Vec3fa*[in->numTimeSteps()];
+      for (size_t i=0; i<in->numTimeSteps(); i++)
+        dnormals[i] = in->dnormals[i].data();
     }
     
     hairs = (ISPCHair*) in->hairs.data();
@@ -297,6 +303,7 @@ namespace embree
     delete[] positions;
     delete[] normals;
     delete[] tangents;
+    delete[] dnormals;
   }
 
   ISPCPointSet::ISPCPointSet (TutorialScene* scene_in, RTCGeometryType type, Ref<SceneGraph::PointSetNode> in)
@@ -504,16 +511,25 @@ namespace embree
     for (unsigned int t=0; t<mesh->numTimeSteps; t++) {
       rtcSetSharedGeometryBuffer(geom, RTC_BUFFER_TYPE_VERTEX, t, RTC_FORMAT_FLOAT4, mesh->positions[t], 0, sizeof(Vec3fa), mesh->numVertices);
     }
+    
     if (mesh->normals) {
       for (unsigned int t=0; t<mesh->numTimeSteps; t++) {
         rtcSetSharedGeometryBuffer(geom, RTC_BUFFER_TYPE_NORMAL, t, RTC_FORMAT_FLOAT3, mesh->normals[t], 0, sizeof(Vec3fa), mesh->numVertices);
       }
     }
+    
     if (mesh->tangents) {
       for (unsigned int t=0; t<mesh->numTimeSteps; t++) {
         rtcSetSharedGeometryBuffer(geom, RTC_BUFFER_TYPE_TANGENT, t, RTC_FORMAT_FLOAT4, mesh->tangents[t], 0, sizeof(Vec3fa), mesh->numVertices);
       }
     }
+
+    if (mesh->dnormals) {
+      for (unsigned int t=0; t<mesh->numTimeSteps; t++) {
+        rtcSetSharedGeometryBuffer(geom, RTC_BUFFER_TYPE_NORMAL_DERIVATIVE, t, RTC_FORMAT_FLOAT3, mesh->dnormals[t], 0, sizeof(Vec3fa), mesh->numVertices);
+      }
+    }
+    
     rtcSetSharedGeometryBuffer(geom, RTC_BUFFER_TYPE_INDEX, 0, RTC_FORMAT_UINT, mesh->hairs, 0, sizeof(ISPCHair), mesh->numHairs);
     if (mesh->type != RTC_GEOMETRY_TYPE_FLAT_LINEAR_CURVE)
       rtcSetGeometryTessellationRate(geom,(float)mesh->tessellation_rate);