+++ doc/src/changelog.md
@@ -1,6 +1,10 @@
 Version History
 ---------------
 
+### New Features in Embree 3.5.0
+-   Changed normal oriented curve definition to fix waving artefacts.
+-   Fixed bounding issue for normal oriented motion blurred curves.
+
 ### New Features in Embree 3.4.0
 -   Added point primitives (spheres, ray-oriented discs, normal-oriented discs).
 -   Fixed crash triggered by scenes with only invalid primitives.